Add tests for depth limit

+++ src/main/java/convex/core/Constants.java
@@ -55,7 +55,7 @@ public class Constants {
 
 	public static final long MAX_SUPPLY = 1000000000000000000L;
 
-	public static final long MAX_DEPTH = 256;
+	public static final int MAX_DEPTH = 256;
 
 	public static final AVector<ACell> INITIAL_GLOBALS = Vectors.of(
 			Constants.INITIAL_TIMESTAMP, 0L, Constants.INITIAL_JUICE_PRICE);
